[MIG] web_widget_one2many_tree_line_duplicate: Migration to 15.0
