+++ web_widget_one2many_tree_line_duplicate/static/src/list/list_renderer.scss
