--- web_widget_one2many_tree_line_duplicate/view/assets.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <template id="assets_backend" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <link
-                rel="stylesheet"
-                href="/web_widget_one2many_tree_line_duplicate/static/src/scss/one2many_tree_line_duplicate.scss"
-            />
-        </xpath>
-        <xpath expr="//script[1]" position="before">
-            <script
-                type="text/javascript"
-                src="/web_widget_one2many_tree_line_duplicate/static/src/js/basic_model.js"
-            />
-            <script
-                type="text/javascript"
-                src="/web_widget_one2many_tree_line_duplicate/static/src/js/one2many_tree_line_duplicate.js"
-            />
-        </xpath>
-    </template>
-</odoo>
